feat(ee): WS5 — relocate proactive-chat impl to ee/src/proactive behind ProactiveService Tag (AC2, #3999)
Physically relocate the 14 proactive-chat implementation modules + the
AnswerMeter SQL impl out of packages/api/src/lib/proactive into
ee/src/proactive. The five admin routes stay in core (EE contributes
Layers, never routes) and reach the relocated logic through two
EE-provided Context.Tags — AnswerMeter (existing) + a new composite
ProactiveService — mirroring the AbuseResponse / MarketplaceVeneer seams.

Core keeps only the declarative surface: the Tag interfaces, Noop layers,
createAnswerMeterTestLayer, the shared types/constants, and the
channel-directory port (hoisted to lib/proactive/types so the core Slack
adapter can implement it without importing @atlas/ee). The runHandler
route (admin-proactive.ts) reaches ProactiveService via runEnterprise.

AC2 only — AC1 (#4064 per-tier entitlement gate) is untouched and still
enforced at the route layer. No core file statically imports @atlas/ee
(check-ee-imports green); the ee-stub build type-checks core standalone.
