+++ ee/src/proactive/__tests__/announcement-coordinator.test.ts
