+++ ee/src/proactive/__tests__/channel-directory.test.ts
