+++ ee/src/proactive/__tests__/workspace-config-loader.test.ts
