+++ ee/src/proactive/announcer-registry.ts
