+++ ee/src/proactive/answer-adapter.ts
