+++ ee/src/proactive/enabled-gate.ts
