+++ ee/src/proactive/user-resolver.ts
