+++ ee/src/proactive/workspace-config-loader.ts
