+++ packages/api/src/api/routes/admin-proactive-public-dataset.ts
@@ -39,15 +39,10 @@ import { runEffect } from "@atlas/api/lib/effect/hono";
 import {
   AuthContext,
   ProactiveGate,
+  ProactiveService,
   RequestContext,
 } from "@atlas/api/lib/effect/services";
 import { hasInternalDB } from "@atlas/api/lib/db/internal";
-import {
-  addEntry,
-  getAllowlist,
-  removeEntry,
-  summarizePublicRefused,
-} from "@atlas/api/lib/proactive/public-dataset";
 import { AuthErrorSchema, ErrorSchema } from "./shared-schemas";
 import {
   createAdminRouter,
@@ -263,10 +258,8 @@ adminProactivePublicDataset.openapi(listEntriesRoute, async (c) =>
         return c.json({ entries: [] }, 200);
       }
 
-      const entries = yield* Effect.tryPromise({
-        try: () => getAllowlist(orgId),
-        catch: (err) => (err instanceof Error ? err : new Error(String(err))),
-      });
+      const proactiveSvc = yield* ProactiveService;
+      const entries = yield* proactiveSvc.getAllowlist(orgId);
       return c.json({ entries }, 200);
     }),
     { label: "list proactive public dataset" },
@@ -302,10 +295,8 @@ adminProactivePublicDataset.openapi(upsertEntryRoute, async (c) =>
       const body = c.req.valid("json");
       const denyMetrics = body.denyMetrics ?? [];
 
-      yield* Effect.tryPromise({
-        try: () => addEntry(orgId, body.entityName, denyMetrics),
-        catch: (err) => (err instanceof Error ? err : new Error(String(err))),
-      });
+      const proactiveSvc = yield* ProactiveService;
+      yield* proactiveSvc.addEntry(orgId, body.entityName, denyMetrics);
 
       log.info(
         {
@@ -368,10 +359,8 @@ adminProactivePublicDataset.openapi(deleteEntryRoute, async (c) =>
         );
       }
 
-      const result = yield* Effect.tryPromise({
-        try: () => removeEntry(orgId, entityName),
-        catch: (err) => (err instanceof Error ? err : new Error(String(err))),
-      });
+      const proactiveSvc = yield* ProactiveService;
+      const result = yield* proactiveSvc.removeEntry(orgId, entityName);
       if (!result.removed) {
         return c.json(
           { error: "not_found" as const, message: "Public dataset entry not found.", requestId },
@@ -420,10 +409,8 @@ adminProactivePublicDataset.openapi(refusedRollupRoute, async (c) =>
       }
 
       const sinceMs = parseSinceMs(c.req.query("since"));
-      const rollup = yield* Effect.tryPromise({
-        try: () => summarizePublicRefused(orgId, sinceMs),
-        catch: (err) => (err instanceof Error ? err : new Error(String(err))),
-      });
+      const proactiveSvc = yield* ProactiveService;
+      const rollup = yield* proactiveSvc.summarizePublicRefused(orgId, sinceMs);
       return c.json({ sinceMs, rollup }, 200);
     }),
     { label: "summarize proactive public refused" },
